feat: updating type-guards to use Zod parsing for more reliable runtime validation (PL-753) (#790)
<!-- You can erase any parts of this template not applicable to your Pull Request. -->

**Fixes or implements PL-753**

### Brief description. What is this change?

<!-- Build up some context for your teammates on the changes made here and potential tradeoffs made and/or highlight any topics for discussion -->

Updated the request type guards in `general-runtime` to use Zod parsing instead. This provides more reliable runtime validation. 

Previously, the `general-runtime` only checke the `type` property of the request object, which was unreliable as the actual data sent to the API may not match the type declaration. By using Zod, we provide complete validation that the incoming request object matches the type.