feat: add adapters for malformed requests coming from react-chat (PL-1037) (#816)
<!-- You can erase any parts of this template not applicable to your Pull Request. -->

**Fixes or implements PL-1037**

### Brief description. What is this change?

<!-- Build up some context for your teammates on the changes made here and potential tradeoffs made and/or highlight any topics for discussion -->

The older versions of the `react-chat` send malformed request objects with `payload: null`. This PR adds adapters on the `general-runtime` to adapt those malformed request objects into ones that actually conform to our defined types.

We opted for this adapter solution rather than changing the request types as this approach conforms to good practices. 
1. An adapter converts bad data into good data, so that `general-runtime` has stronger guarantees of the incoming data.
2. Changing the types to adhere to bad data is bad practice as it leads to gradual deterioration of the reliability of the types