feat: add trace time property (ENG-172) (#866)
add a `time` property with `Date.now()` for all traces.

Right now the time property is optional in `BaseTraceFrame`,  however for the purpose of the PR I intentionally modified my local type to have `time` to be required and fixed all resulting type and build errors. This way we can remove the optional field and have it fully be required.

Reference my comments for more details. The majority of traces just use `runtime.trace.addTrace` so I made it automatically assign the time there, and have the time field optional so other places can be less verbose.

Co-authored-by: Tyler <tylerhan97@gmail.com>