feat: change script condition execution to expect the function body instead of whole function definition (PL-1273)  (#872)
<!-- You can erase any parts of this template not applicable to your Pull Request. -->

**Fixes or implements PL-1273**

### Brief description. What is this change?

<!-- Build up some context for your teammates on the changes made here and potential tradeoffs made and/or highlight any topics for discussion -->

Product discussion came to the conclusion that it's simpler to have the script condition contain only the function body, that is, the code will look like it supports top-level return. This is how it's done on the JS step. We will not be showing the first and last line of the function body, e.g, the declaration `function condition({ variables })`. 

<img width="770" alt="Screenshot 2024-07-10 at 3 42 16 PM" src="https://github.com/voiceflow/general-runtime/assets/32404412/f1ebb963-d03e-44de-9c15-f89d302bef4d">

This PR reworks the code sandbox to expect the function body, instead of the complete function definition. 

The image below shows how the script condition previously worked.

<img width="761" alt="Screenshot 2024-07-10 at 3 45 28 PM" src="https://github.com/voiceflow/general-runtime/assets/32404412/510493af-9618-47ff-81d2-f3eac38c3710">