fix: fixing crash due to empty variants (COR-2571) (#870)
<!-- You can erase any parts of this template not applicable to your Pull Request. -->

**Fixes or implements COR-2571**

### Brief description. What is this change?

<!-- Build up some context for your teammates on the changes made here and potential tradeoffs made and/or highlight any topics for discussion -->

The message step handler crashes when all of the unconditioned variants have no text.

This is because the compiler (the `realtime`?) is stripping out variants with empty text. The `general-runtime` assumed that at least one variant would exist on a message resource, because the UI ensures there is always an editor for the unconditioned variant resource, but this requirement become invalid due to the compiler's stripping. 

So when executing this incorrect message resource, we attempt to sample from the list of unconditioned variants (which is `[]`) and then access the `.variant` property, `unconditionedVariants[randomIndex].variant`, causing an exception. 

This PR and a PR on `general-service` fixes the issue. The compiler no longer strips variants with empty text and the `general-runtime` will not output a trace with empty text.